+++ src/args/arg_builder/positional.rs
@@ -17,10 +17,10 @@ pub struct PosBuilder<'n, 'e> {
     pub requires: Option<Vec<&'e str>>,
     pub blacklist: Option<Vec<&'e str>>,
     pub possible_vals: Option<Vec<&'e str>>,
-    pub index: u8,
-    pub num_vals: Option<u8>,
-    pub max_vals: Option<u8>,
-    pub min_vals: Option<u8>,
+    pub index: u64,
+    pub num_vals: Option<u64>,
+    pub max_vals: Option<u64>,
+    pub min_vals: Option<u64>,
     pub val_names: Option<VecMap<&'e str>>,
     pub validator: Option<Rc<Fn(String) -> StdResult<(), String>>>,
     pub overrides: Option<Vec<&'e str>>,
@@ -50,7 +50,7 @@ impl<'n, 'e> Default for PosBuilder<'n, 'e> {
 }
 
 impl<'n, 'e> PosBuilder<'n, 'e> {
-    pub fn new(name: &'n str, idx: u8) -> Self {
+    pub fn new(name: &'n str, idx: u64) -> Self {
         PosBuilder {
             name: name,
             index: idx,
@@ -58,7 +58,7 @@ impl<'n, 'e> PosBuilder<'n, 'e> {
         }
     }
 
-    pub fn from_arg(a: &Arg<'n, 'e>, idx: u8, reqs: &mut Vec<&'e str>) -> Self {
+    pub fn from_arg(a: &Arg<'n, 'e>, idx: u64, reqs: &mut Vec<&'e str>) -> Self {
         assert!(a.short.is_none() || a.long.is_none(),
             format!("Argument \"{}\" has conflicting requirements, both index() and short(), \
                 or long(), were supplied", a.name));
@@ -150,13 +150,13 @@ impl<'n, 'e> AnyArg<'n, 'e> for PosBuilder<'n, 'e> {
     fn is_set(&self, s: ArgSettings) -> bool { self.settings.is_set(s) }
     fn set(&mut self, s: ArgSettings) { self.settings.set(s) }
     fn has_switch(&self) -> bool { false }
-    fn max_vals(&self) -> Option<u8> { self.max_vals }
-    fn num_vals(&self) -> Option<u8> { self.num_vals }
+    fn max_vals(&self) -> Option<u64> { self.max_vals }
+    fn num_vals(&self) -> Option<u64> { self.num_vals }
     fn possible_vals(&self) -> Option<&[&'e str]> { self.possible_vals.as_ref().map(|o| &o[..]) }
     fn validator(&self) -> Option<&Rc<Fn(String) -> StdResult<(), String>>> {
         self.validator.as_ref()
     }
-    fn min_vals(&self) -> Option<u8> { self.min_vals }
+    fn min_vals(&self) -> Option<u64> { self.min_vals }
     fn short(&self) -> Option<char> { None }
     fn long(&self) -> Option<&'e str> { None }
     fn val_delim(&self) -> Option<char> { self.val_delim }
